Add precomputed linear kernel + test
kernel::linear_precomputed is not included in the default kernel.hpp header, as
it's not intended for actual use but rather as the simplest test case for a
precomputed kernel.

Using some SFINAE magic to specialize the behavior of model::operator() for
precomputed kernels -- in that case the test data needs to be "kernelized"
before prediction can take place. One subtlety here: need an additional dummy
template parameter in one of the "overloads" to distingish their template
signature. The compiler does not reason that both overloads are actually
mutually exclusive and complains about a potential overload clash otherwise. In
the future, one would be better of to switch the behavior by the use of C++17's
`if constexpr` construct, but I wanted to forego introducing a C++17 dependency
just yet.