Introspect into linear model to obtain hyperplane coeffs + test
Changed introspection inheritance: introspection_policy becomes
introspective_model and inherits _from_ model, rather than the other way around.
This way the introspective_model ctor can perform post-processing, _after_ the
training of the model took place. This also gives access to the model w/o the
need for a self-reference to be passed to the ctor.