Enable precomputed problems to use arbitrary containers for samples
Not need to store original sample data in libsvm's awkward node structure. We
can use any container for that, e.g. std::vector and have the kernel operate on
those. This allows to potentially avoid copy-on-sample by exploiting move
semantics. It also may be beneficially for large sample sizes as it saves on
space. Last, it may also avoid breaking vectorization of the kernelization loop.

For now, dataset (encapsulating libsvm's node structure) is still used in
precomputed_linear kernel.