Split serialization in model_ and problem_serializer
This allows to use the problem_serializer to save and restore problems to and
from checkpoints when the simulation is interrupted before the model is
constructed. In that case problems need to save their data either way, even if
the Kernel is not precomputed. An optional boolean flag can be passed to the
problem_serializer ctor to indicate is the sample data should be skipped.