Unsymmetrize coefficients
Use symmetrized coefficient array but assign values for all rank! indices
corresponding to the (rank)-tuple permutations which are bound to be equivalent.
Doing so restores the full (unsymmetrized) coefficient tensor which eases
interpretability while maintaining the significant (rank!)^2-speedup in the
introspection step. Failing to do so would have been prohibitive for order
parameter tensors of rank 4 and up.