Add config policy for resolution of sites (rank 1)
This is limited to rank 1 (Heisenberg, FM & AFM) for now and resolved every
site. This can be extended in the future to higher rank (also the size of the
configuration vector may quickly become prohibitively large) and unit cell
policies where only some sites are resolved (w.r.t. their location in a super
cell).