Move results stuff in its own namespace
The contraction class for results previously conflicted with the contraction
class used in the contraction analysis of the result. The latter was
consequently moved in a "detail" namespace as a quick fix. This is no longer
needed.

I looked into replacing the contractions to generate the result tensor with
those used in the analysis to avoid similar code. For component indices this
would work (you'd specify the "x" vector that multiplies the contraction matrix,
but there's currently no way to identify blocks, so we leave it as-is for now.