Support multi-classification in the test program
The line_scan sweep policy is used to generalize the sweep through the phase
diagram. The parameter file can specify start and end point of the sweep by the
sections [test.a] and [test.b].

All decision function values are measured, as well as optional additional
observables for true order parameters. Their number is now variable and
simulation base classes need to provide a order_param_names() *member* function,
rather than a static single string variable. The amount may also vary with phase
point.

The SVM decision functions are no longer matched with the true order parameter
as this is now ambiguous. They are instead rescaled to an interval of unit
length, but *not* shifted, meaning that the zero-crossing corresponds to the
decision boundary.

Updated SVM submodule to include permutation of classifiers and labels to
guarantee a stable ordering w.r.t. to the Label's operator<.