Guard against construction of iterator w/ past-the-end unitcell iter
When a past-the-end unitcell iterator is passed to the basic_iterator ctor, the
basis index is set to zero. This is important for the newly constructed
basic_iterator to be regarded as a past-the-end iterator itself, i.e. --it being
dereferenciable and comparing equal to end().

This is particularly relevant with open boundary conditions where the unitcell
iterators passed to the ctors in nearest_neighbors may be past-the-end because
the Bravais lattice boundary is reached.