Update ising code to the new SVM interface; add ising_config_policy
Use a config policy (derived from monomial_config_policy) for the ising code as
well.

Retired the old ising-coeffs program in favor of one derived from the generic
coeffs program. This was done for consistency. We loose the ability to calculate
the coefficient matrix as a function of distance (exploiting translational symm;
see PonteMelko17). Higher-rank monomial kernels can be used (which is pointless
of course for the Ising model, but still).