Restrict the single_flip update on spin_O3 systems to a "cap"
At low temperatures, drastic changes of a single O(3) spin are most likely
rejected pretty much all the time. To propose primarily updates which have a
decent change of being accepted, we can draw the proposed new spin from a "cap"
around the old spin.

The new spins are generated as follows:
1.) A orthogonal rotation matrix is constructed from the polar coords of the old
spin. This matrix (R) will essentially rotate the z-axis onto the old
spin (prior to the update).
2.) A new spin is generated. Unlike before, its z-component is not drawn
uniformly from [-1;1], but from [cos(theta_0);1]. Thus, this new spin will
deviate from the z-axis by a polar angle of at most theta_0.
3.) The matrix R is applied to the new spin to yield one "close" to the old
spin.

cos(theta_0) becomes a new parameter to control this. It can be set from the
parameters (cf. parent commit) as "update.single_flip.cos_theta_0". The default
value of -1 corresponds to the isotropic choice that was previously implemented.

In testing, one typically observes the following behaviour:
As cos_theta_0 is increased, the acceptance rate goes up and the
autocorrelations go down for a while before they start to grow again as
cos_theta_0 approaches 1. In the latter case, we are precluding updates which
would have had a decent chance of being accepted, but instead carry out a lot of
tiny steps which are accepted most of the time but don't alter the configuration
enough. The optimal autocorrelation is typically reached for acceptance rates of
10%-50%.

Benchmarking reveals that the additional overhead due to the more complicated
update (construction of the matrix, matrix-vector multiplication) is entirely
negligible and thus this update mode is enabled even if only isotropic
proposals (the default) are used.