Construction of sweep policy outsourced
... to `sweep::from_parameters<>(alps::params const&)`. This was
previously done (in a somewhat bulky) lambda in the test_adapter ctor.

This function is now also called in the segregate_phases program. This,
together with the fixed size of the sweep policy (cf. prev commit),
allows segregate_phases to work with arbitrary sweep policies and not
just `sweep::grid`.

(Note: This change was made with particular regard for
`sweep::nonuniform_grid`. Be aware though that non-uniform spacing of
the phase points will in principle bias the choice of the threshold
bias rho_c.)