Disable comparison w/ true PD in segregate_phases
This has D2h phase diagram (gauge project) as a reference hard-coded.
Since this was done merely to research (and validate) the threshold
bias criterion and no reference phase diagram exists for models with
unknown physics, this is phased-out for now. This will allow
segregate-phases to become applicable to other projects.