Make sweep policies sized
Originally sweep policies were designed to yield phase space points
indefinitely and the `sweep.N` parameter controlled the total amount
that should (at least) be drawn. This has been revised to where the
individual sweep policies store this number and it can be queried using
the (virtual) size() function that is now part of the sweep::policy
interface.