Store (and apply) masks to graphs in segregate-ph.
The connectivity of the graph resulting from the current invocation of
the segregate-phases program is stored in the file "mask.txt". For each
tentative graph edge (corresponding to the SVM classifiers among
points), 0 indicates that the graph does *not* have an edge there, while
1 means that it does.

Using the flag `--mask <file>`, one can "import" the mask from another
set of classifiers and it will act as a "mask" to the current
classifier's graph, putting an edge in the graph only if there would
have been an edge given the current classifier anyway *and* if there was
also an edge for the classifier used to calculate the mask. Thus, a
logical AND of the two graphs is performed. Note that the resulting
graph is used to write mask.txt for the current invocation, which means
that repeated invocations with the previous invocation's mask will
logically-AND multiple graphs.

This can be used to combine the classification power of different
parameter sets; in particular, this can be used to combine data from
multiple ranks.