Add config_serialization header
This provides the ability to copy a simulation's configuration in a
linear container and restore it from there. This will be used in HDF5
serialization of configurations.

The generic implementation in this header assumes that configuration
types support Container-type operations (begin and end member functions,
returning iterator that dereferentiate to double, size member function).
Client code which have configurations that do not support these should
expose a template specialization of config_serializer<> instead to
provide custom code.