Add procrastination_adapter & use it for learning
The procrastination_adapter is a drop-in replacement for the
training_adapter. But instead of applying the kernel (through confpol)
eagerly when a new sample is taken, the procrastination_adapter stores
the full configuration that has been sampled and applies the kernel
lazily only when surrender_problem is called. Hence the name: instead
of learning continuously throughout the simulation / semester, the
procrastinator will accumulate material and start learning only just
before the SVM optimization / exam. ;-)

This approach has several advantages:
[+] The configuration policy is not fixed during the sampling stage.
    The user can resume a simulation from the *.clone.h5 file,
    --skip-sampling, and override parameters such as rank, cluster
    choice, etc.
[+] This avoids running the same simulations again just to learn an SVM
    for given confpol parameters. This is particularly advantageous for
    expensive MC simulations.
[+] The configurations are stored in the *.clone.h5 in a space-efficient
    binary format and can take advantage of HDF5's built-in gzip
    compression.

... but comes with an obvious downside:
[-] The resulting *.clone.h5 file may be quite large, especially for
    large lattices.