Add flag --skip-svm to learn program
This allows to only run the MC simulation and sample, but skip the
subsequent SVM optimization. This may be useful in conjunction with
lazy config policy evaluation and/or for grid classifier runs with a
large number of grid points where the SVM optimization can take very
long, such that a separate cluster job or different job requirements
may be prudent.