Restrict connectivity range in segregate-phases
Using the --radius | -R flag, a maximum range of edges to consider for
adding to the graph can be specified. The idea is that short-range
connections are sufficient to establish connected components, while
long-range edges to not add any new information and may potentially have
a higher chance of introducing misleading connections, establishing
connections between far-away regions in the phase diagram.