Use weighted graph analysis based on bias value
Instead of making a binary choice based on | |rho| - 1 | > rho_c for
whether or not to add an edge to the graph, we now opt to constructed a
weighted graph where the weight is determined from the value of the bias
parameter through some weighting function, e.g. an (inverse) Gaussian,
where the weight is in [0, 1].

The Laplacian of a weighted graph is constructed analogously to that of
an unweighted one, except edges enter the degree and adjacency matrix
according to their weight.

In principle, all possible edges (at least those qualified by the range
check, cf. a53ba325) appear in the graph, albeit potentially with a very
small weight. As a result, the zero eigenvalue is almost never going to
be truly degenerate.

The output file "edges.txt" now follows a different format, putting
"start" and "end" vertex coordinates in the same line, followed by its
weight. FTR, this can be used to plot the graph in a way where the line
width corresponds to the weight:

gnuplot> set for [i=1:101] style arrow i linewidth i/100. nohead
gnuplot> plot 'rank2/edges.txt' \
         using 1:2:($3-$1):($4-$2):(int($5*100)+1) \
         with vectors arrowstyle variable