Calculate RMS bias deviation
In *-segregate-phases, the bias distribution is examined before anything
else happens. We calculate the "standard deviation" w.r.t. to an assumed
peak value of rho=1.

This is motivated by an desire to automate the choice of rho_c and give
an unbiased prescription who to do so. For the "box" weight function,
the old approach was to choose rho_c such that it eliminates the most
amount of graph edges before the graph was to become disconnected, i.e.
the Laplacian's zero eigenvalue would become degenerate. This is no
longer feasible when a continuous weighting function is used. The choice
of a gaussian weighting function motivated us to consider the standard
deviation, i.e. fit the weighting function to the bias data.

However, the bias distribution is not really symmetric (being lower-
bound by 0) and has a long tail. Thus, the RMS (root mean square) of the
bias deviation from 1 is generally too large to make for a suitable
rho_c. In the future, a fit of a long-tail distribution (Levy alpha) may
be a more prudent choice, both to estimate rho_c, but also as weighting
function.