Lorentzian bias distribution
To address the long tails of the distribution of bias values (cf. prev.
commit msg), a Lorentzian distribution has been added. The width (Gamma)
of the Lorentzian is given by rho_c (--rhoc | -r) but its default value
is determined from half of the interquartile spacing of the bias values.
(The Gaussian distribution still uses the sample variance.)