Add ability to combine graphs based on weight
This restores the feature that was temporarily removed in 54126417.
The user can specify one or multiple "rho.txt"-type files from other
classifications as positional arguments on the command line, e.g.

$ ./gauge-segregate-phases D3h_rank2.out.h5 -w gaussian ../rank3/rho.txt

The *weights* determined when writing the auxiliary rho.txt files are
multiplied with those obtained for the current classification. I.e. the
selected weighting function and rho_c parameter DO NOT AFFECT the
auxiliary graphs but only the current one. On the contrary, the range
option (--radius | -R) of the current invocation override any different
choice made in the other classifications.

To obtain the original behavior for combining graphs using logical AND,
use this feature in combination with the "box" weighting function.