Output "mask" in mask.txt based on threshold
Binary label (0 or 1) indicating if eigenvector entry is larger than a
given threshold value (--threshold | -t) is written. This can be
used as a mask in upcoming commits to restrict the graph analysis to a
subset of full graph. This may be used as an alternative means to
decompose the graph in successively (-> "hierarchical clustering").