Apply mask to perform graph analysis on subset
Using the flag --mask | -m, one can specify a "mask.txt" file from a
previous graph analysis on the same phase space grid and restrict this
graph analysis on the subgraph defined by the mask. This is useful in
hierarchical clustering analyses.

Additional, a mask can be inverted on-the-fly by specifying
--invert-mask .