Measure & output decision func variance in test
Added <d^2> as an observable in the test_adapter and calculate the
decision function's variance <d^2> - <d>^2 in the test program.

The decision function's variance may be a quantity akin to the Binder
cumulant and thus may be useful to pin-point tentative transitions based
on the order parameter / decision function curve.