Add --masked-value option to segregate-phases
When a mask is applied (--mask | -m) on the graph analysis, the
corresponding phase points are removed from consideration and from
output. Using the --masked-value option, the latter can be reversed by
assigning the masked phase points a fictitious value rather than
omitting them entirely. This is useful when plotting the Fiedler vector
"with image" in gnuplot as this plot style requires a regular grid
without any omissions.