Add flag --infinite-temperature in learn program
To compare phases to the "infinite temperature" state where all spins
are randomly oriented, a new flag --infinite-temperature, -i has been
introduced to the learn program. When specified, an additional "fake"
phase space point is "sampled" were the configurations are perfectly
random.

To achieve this, the SVM interface for client codes has been expanded to
include a random_configuration() function. This is a breaking change as
client codes have to implement this from now on.

This feature is intended to be used with the fixed_from_cycle classifier
which has been made aware of the "fake" infinity phase point and will
classifier those samples with a distinct label.