Reimplement mpi::environment to be thread-aware
Using the dispatcher in a std::thread led to occasional segfaults and
aborts. This is because MPI is not thread-safe by default.

alps::mpi::environment initializes with MPI_Init. There exists however
an alternative initialization with MPI_Init_thread which requests a
certain level of thread-safety from the MPI implementation.

For now, we will request full thread safety. When this is unavailable on
production systems, we may instead opt for the "serialized" option which
does allow MPI calls from different threads, but requires us to ensure
they are not concurrent.