Avoid spinning on dispatcher thread
Most of the time, the dispatcher thread waits for one of the groups to
report idle. The default MPI impl will spin while waiting, leading to
100% CPU utilization for the dispatcher thread which however shares its
resources with the rank 0 worker. Instead, now using non-blocking MPI
routines with a custom spin loop which puts the dispatcher to sleep for
100 ms. This way, the worker is not outmaneuvered by the dispatcher
thread which adding negligible latency to the batch dispatch.