Use MPI mutex to guard HDF5 ops in dispatcher
Previously, the dispatcher used a custom ring lock. This has been
retired in favor of the MPI-based mutex that is already being used to
guard the HDF5 operations when writing the results to the test file. By
using the same mutex, we also make sure that both these parts (results
and dispatcher) respect each other. This eliminates the need for some
MPI barriers.