Collect data from clones working on the same sim
Introduced `embarrassing_adapter` which is derived from ALPSCore's
`mcmpiadapter` to achieve parallelization of batches in an
embarrassingly parallel way. It extends the vanilla `mcmpiadapter` by
interface function allowing to change the underlying MPI communicator
(which may be necessary if the batch is smaller than the group itself
such that only some clones are working on a valid point) and to query
the number of different phase points that results have to be collected
for.

The general interface has been design with the feature generalization in
mind that the `embarrassing_adapter` might be replaced by one that
supports parallel tempering.