Generalized batching of phase points
The simulation adapter (for now only the `embarrassing_adapter`) defines
a `batcher` type which is a parameter-initializable functor which maps a
container of phase_points to a container of batches (which are
themselves containers of phase_points).

The `embarrassing_adapter`'s batcher duplicates each phase points a
given number of times. This replaces the `--repeat` flag previously
introduced in the sample and test programs. The corresponding parameter
is `batch.n_parallel`. If the special value zero is passed, the batches
will be sized such that they achieve the optimal utilization of the MPI
world.