Fix bug: ISING define name collision (frustmag)
The ISING preprocessor variable definition used to select the ising
model in the frustmag client code clashed with that which selects the
ising client code. The former has been renamed to ISING_HAMILTONIAN
(along with HEISENBERG_HAMILTONIAN) to avoid this. This change does not
affect the user-facing values of the HAMILTONIAN CMake variable.