Add pt_adapter with iso_batcher
This adapter is designed to support parallel tempering (PT). Using the
iso_batcher, phase_points are batched such that all points which differ
only in one ("temperature") parameter, though other batchers are
possible in principle.

This is the first prototype: for testing the pt_adapter is used by
default; proper collection of measurements is not yet implemented.