Merge measurements within each process
Each MPI process has a repository of measurements for each phase_point
it "visited" during its PT "random walk" (the latter not being
implemented yet).

When the measurements are finally collected, the measurements in each
"slice" need to be merged. Currently, this is only implemented within
each process and needs to be extended to a collective operation in the
future.