Reset measurements in reset_sweeps; interface ch'd
SVM interface has changed: update_phase_point(...) is now supposed to
return a bool, indicating whether the point has indeed changed (i.e.
fresh thermalization is in order). It should *no longer* call
reset_sweeps itself. This is because it is intended to be used in PT to
update the current phase space point w/o resetting the progress of the
simulation.

As a consequence, the test_adapter is also no longer allowed to reset
the simulation's measurements member; this is now done in reset_sweeps.
Analogously it befalls to the pt_adapter to clear the measurements for
all the slices in reset_sweeps. I.e. this would be the case when the PT
*batch* is changed, but not when the PT update issues a swap of phase
space points.