Use pt_adapter as base class of frustmag sim
Follow-up to prev commit. The pt.hpp is gone; the `negotiate_update` has
been migrated to the pt_adapter itself. Its signature is more compact
because the pt_adapter already has access to a lot of the relevant info.

The reference to the `pt_adapter` has to be passed (rather ugly) to the
`parallel_tempering` update. Meta-programming is used to fiddle the
reference through the update mux. Not ideal, perhaps use ctor (template)
argument for all updates instead and ignore if not needed. On the other
hand, no rebind_communicator magic is needed anymore.

Since plain references cannot be rebound, the measurements object is
exposed (as a protected member) as a std::reference_wrapper.
Consequently, the derived simulation classes have to call `get()` on it.
This change has only been done for the frustmag code for now, breaking
the interface for the ising and gauge codes. Perhaps, we should use a
member function `measurements()` instead. This is more elegant and is
arguably better practice than exposing member variables in the first
place. However, this choice has been made by ALPSCore devs and changing
it breaks the interface. All other things being equal, the current
solution also breaks the interface, therefore this is probably still an
acceptable price to pay.