measurements() member function shadows member var
The change described in the commit message of 489a6f276a has been
implemented. Client codes have to replace any usage of the protected
`measurements` member variable (of `alps::mcbase`) with the
`measurements()` member function. Failing to do so generates errors
(because the function shadows the variable) thus accidental misuse is
precluded.