Replace set_pta call with variadic ctor args
This is the mechanism used to supply the `parallel_tempering` update in
the frustmag sim with a reference to the `pt_adapter` to invoke
`negotiate_update`. The previous solution was ugly (required tons of
metaprogramming to fiddle through the update muxer). We now pass a
reference to the simulation as an addition ctor argument to all updates.
The update ctors can choose to bind these additional arguments or absorb
them in a variadic argument pack.