Collective merge of results across all PT ranks
Previously, only the results for the *current* phase point on any given
rank that were also measured on that rank have been collected. Now, we
go through all the ranks and each rank collects results for its
*current* phase point from *all* other ranks which have also accumulated
measurements for that phase point.

`mpi::split_communicator` is used to create a sub-communicator of only
those ranks which have accumulated a finite amount of measurements. This
is a clear improvement over the behavior of `alps::mcmpiadapter` which
throws in case not all ranks have measurements for a given observable.