Fix bug: reinstate empty accumulators upon load
ALPSCore does not save (and therefore load) empty accumulators in/from a
measurement set stored in HDF5 format. This is problematic for example
when restoring from a simulation that has not yet left its
thermalization phase. Since we occasionally switch phase space points
and thus enter into a new thermalization phase, this could happen at any
point, not just the early stages. This was the problem reported in the
previous commit message.

The workaround is similar to the procedure used in
`pt_adapter::update_phase_point` to duplicate accumulator sets for each
slice with preregistered accumulators. The old current slice measurement
object is used as a prototype to reinstate any empty observables in the
restored measurements.