Keep track of PT slice permutation
Before, the PT manager always proposed updates between adjacent *rank*.
This is of course nonsensical since the physical phase space points will
be shuffled after a good amount of successful PT updates. Thus, this
scheme ended up proposing updates between points which are far apart
which are almost certainly rejected.

The PT manager now keeps track of this permutation and inverts it to
propose updates which are physically adjacent. At the beginning of each
new simulation (when reset_sweeps is called) the permutation is
initialized with identity (iota). After each successful update, the
permutation is updated with the additional transposition. The
permutation is also stored and restored. This is crucial to avoid mixing
measurements from the wrong points.

This ability necessitated two changes in the MPI communication protocol:
* The primary rank (the one initiating the update) is sent two
  adjacent ranks to choose from [if possible; invalid choices (boundary
  or unavailable due to an ongoing PT update) are coded as -1]. If
  ambiguous, the primary rank informs the manager of its choice.
* The primary rank has to inform the manager of the result of an update,
  such that the manager can apply the additional transposition. As a
  consequence, the manager can no longer mark both update partners as
  "available" directly after it has communicated to the secondary rank
  its partner; rather, the pair has to be kept off limits such that the
  transpositions remain in-order, as those do not commute.
  This is achieved by two additional `query_types`, `acceptance` and
  `rejection`.