Improvements to `main()` logic in sample / test
These changes mainly reflect the changes made to the `pt_adapter` in the
previous commit. Particularly, when restoring from a checkpoint, the
stored phase space point does *not* need to coincide with the
slice_point that the dispatcher would have assigned; due to PT it may be
permuted.

Also, in the sample program, `reset_sweeps` should be called if the
simulation is *not* restored from a checkpoint.