Fix MPI bug: acknowledge receipt of update status
This addresses a rare deadlock condition which could arise when the
secondary rank in a PT update received its update status from the
primary rank and subsequently "moved on", *before* the manager has
processed the same update status. In that case, it was possible that the
(former) secondary rank filed a new status request with the manager and
got the `secondary` status back, expecting that it was to engage again
with the same primary rank in a new update, when in fact there was no
primary update ongoing.

This problem is resolved by having the manager acknowledge the receipt
of the update's acceptance status and clearing of the participants
statuses to the (former) secondary rank. It has to wait until it
received this acknowledgment from the manager before it can move on.