Keep track of the "upness" of PT ranks and measure
Upness refers to the state of having visited the lowest slice prior to
visiting the highest slice, hence being on the upward leg of a
roundtrip. This is kept track of by the PT manager for all ranks and
stored in an observable "Upness". The entries in the vector are permuted
such that the elements are arranged by slice, rather than by rank.

Because the manager has to get access to the measurement object,
`manage` has been made a member function.