Change manager idle time to 1 us
Testing revealed the previous time (100 ms) to be too long and stall due
to a "busy" manager reduced efficiency. The new value is probably
smaller than what the system allows; `sleep_for` is allowed to sleep for
longer than the time provided. Testing suggests the actual time is more
between 10 and 100 us on my system.

The reduced time does put some extra load on the rank-0 process with CPU
usage averaging between 105% and 125% depending on PT update frequency.
However, this did not slow down the rank-0 worker significantly compared
to its "co-workers" and is definitely worth the speed up of the manager.

On the other hand, not having any sleep for the manager is
counterproductive as the manager will choke the rank-0 worker which
hurts the productivity of the entire team (and is probably not good for
morale).