Add a `log_scan` sweep policy
Allows for the generation of points on a logarithmic scale. Using the
parameters "log1", "log2", ... the use can specify if the axis 1, 2, ...
is to be considered a log scale; by default all axes are considered log.

For example, for a 2d phase space, the policy would generate N equi-
distant points on a line in the log-log plot, i.e. satisfying a power
law relationship. On the other hand, if "log1" is set to false, the
points lie on a line in a lin-log plot, i.e. on an exponential curve in
the corresponding lin-lin plot.