Filter invalid phase labels from the SVM problem
This allows to discard certain regions of phase space. E.g., a phase
diagram classifier may not cover the entirety of the phase space. This
this situation, the unclassifiable data are assigned the "None" label
and filtered from the problem before optimization commences.

Updated SVM wrapper with upstream filter functionality.