Add sweep_grid sweep policy
This sweep policy spans a grid of points in phase diagram space, similar
to `grid` and `nonuniform_grid`. The difference is that whereas `grid`
uses equidistant points along its axes and `nonuniform_grid` uses
user-defined points along its axes, `sweep_grid` defines separate sweep
policies to generate the points along its axes.

That is, the function of a `grid` could be reproduced by combining
`sweep_grid` with `line_scan` policies along the axes, and the function
of a `nonuniform_grid` could be reproduced by combining `sweep_grid`
with `cycle` policies.

The downside of this approach is that the amount of parameters that need
to be defined explodes a bit. To avoid infinite recursion (and because
it's nonsensical), the axis sweeps of a `sweep_grid` policy cannot be
`sweep_grid`s themselves.