Fix bug: do not measure in test_adapter once done
In general it would be fine to keep on sampling when
fraction_completed() exceeds one to gather additional statistics.
However this leads to a bias when the update function ceases to actually
update, measuring the same sample over and over again.
One can also make the argument, that all phase points should accumulate
the exact same amount of statistics in order for variances and mean to
be consistent across the board. For these reasons, no additional samples
are acquired any more after fraction_completed() exceeds 1.