Output variance also for misc. order params
For each of the "order parameters" specified by sim.order_param_names(),
also check of its square is being measured. If so, it too is added as a
column to the output text file and also the variance
(mean of square - square of mean) is calculated (with error propagation
through resampling) and also added as a column.

To check if the square of an observable is measured, "^2" is appended to
its name, unless that name ends in a power, in which case it is doubled.
For example, if "A" and "A^2" are in order_param_names() and "A", "A^2",
and "A^4" are being measured, the output file will include columns for
A, var(A), A^2, var(A^2), and A^4. If "A^4" was absent from the
measurements, only A, var(A), and A^2 would be output.