Fix bug: pt_adapter::load failed if not subset
If the restored slice_measurements contained accumulators which are not
defined in the prototype, the previous (more "elegant") implementation
meight have thrown (depending on lexicographic ordering of accumulator
names). The new impl does allow for additional accumulators in the
restored slice_measurements. This situation may arise of subclasses
define additional variables (as is the case for the test_adapter).