Skip sample is configuration is empty
This enables simulation classes to intentionally return an "empty"
configuration in cases where no samples are desired but the simulation
is merely run as part of a larger PT set.

"Emptiness" is established by calling empty() on the configuration.
Since not all configuration type may implement empty(), the comparison
`size() == 0` is a compile-time fallback. If no `size()` is provided,
the distance between begin and end iterators is used.