Use stored merged_parameters in *-segregate-phases
Rather than locating the clone files based on the `merge` parameter and
constructing the sweep policies of the merged clones directly from their
parameters, we now use the `merged_parameters` which are stored within
the *.out.h5. This way, the graph analysis is yet again dependent on
only the *.out.h5, meaning that it may be performed locally from the
downloaded out file, while the much larges clones remain on the cluster.